Add svg object to configuration
The compiled sprite does not have style and dimension attributes in the <svg> root element so that it can be embbed in a project document without affecting layout.

To fix, we've added the svg object  that allows us to set attributes. Adding width, height and absolute positioning allows us to embed the inline sprite in our project html without affecting other surrounding elements.

Also removed the XML declaration and DOCTYPE declaration because it doesn't seem necessary to have.